Create and modify release bundles
