+++ distribution/services/createbundle.go
@@ -14,69 +14,33 @@ import (
 	"github.com/jfrog/jfrog-client-go/utils/log"
 )
 
-type ReleaseNotesSyntax string
-
-const (
-	Markdown  ReleaseNotesSyntax = "markdown"
-	Asciidoc                     = "asciidoc"
-	PlainText                    = "plain_text"
-)
-
 type CreateReleaseBundleService struct {
-	client        *rthttpclient.ArtifactoryHttpClient
-	DistDetails   auth.CommonDetails
-	DryRun        bool
-	GpgPassphrase string
+	UpdateReleaseBundleService
 }
 
 func NewCreateReleseBundleService(client *rthttpclient.ArtifactoryHttpClient) *CreateReleaseBundleService {
-	return &CreateReleaseBundleService{client: client}
+	return &CreateReleaseBundleService{UpdateReleaseBundleService{client: client}}
 }
 
-func (ps *CreateReleaseBundleService) GetDistDetails() auth.CommonDetails {
-	return ps.DistDetails
+func (cbs *CreateReleaseBundleService) GetDistDetails() auth.CommonDetails {
+	return cbs.DistDetails
 }
 
-func (cbs *CreateReleaseBundleService) CreateReleaseBundle(createBundleParams CreateReleaseBundleParams) error {
-	var bundleQueries []BundleQuery
-	// Create release bundle queries
-	for _, specFile := range createBundleParams.SpecFiles {
-		if specFile.GetSpecType() != artifactoryUtils.AQL {
-			query, err := artifactoryUtils.CreateAqlBodyForSpecWithPattern(specFile)
-			if err != nil {
-				return err
-			}
-			specFile.Aql = artifactoryUtils.Aql{ItemsFind: query}
-			aql := artifactoryUtils.BuildQueryFromSpecFile(specFile, artifactoryUtils.NONE)
-			bundleQueries = append(bundleQueries, BundleQuery{Aql: aql})
-		}
+func (cbs *CreateReleaseBundleService) CreateReleaseBundle(createBundleParams CreateUpdateReleaseBundleParams) error {
+	CreateUpdateReleaseBundleBody, err := CreateBundleBody(createBundleParams, cbs.DryRun)
+	if err != nil {
+		return err
 	}
-
-	// Create release bundle struct
-	releaseBundle := &ReleaseBundleBody{
+	createReleaseBundleBody := &CreateReleaseBundleBody{
 		Name:              createBundleParams.Name,
 		Version:           createBundleParams.Version,
-		DryRun:            cbs.DryRun,
-		SignImmediately:   createBundleParams.SignImmediately,
-		StoringRepository: createBundleParams.StoringRepository,
-		Description:       createBundleParams.Description,
-		BundleSpec: BundleSpec{
-			Queries: bundleQueries,
-		},
+		ReleaseBundleBody: *CreateUpdateReleaseBundleBody,
 	}
 
-	// Add relese notes if needed
-	if createBundleParams.ReleaseNotes != "" {
-		releaseBundle.ReleaseNotes = &ReleaseNotes{
-			Syntax:  createBundleParams.ReleaseNotesSyntax,
-			Content: createBundleParams.ReleaseNotes,
-		}
-	}
-
-	return cbs.execCreateReleaseBundle(releaseBundle)
+	return cbs.execCreateReleaseBundle(createReleaseBundleBody)
 }
 
-func (cbs *CreateReleaseBundleService) execCreateReleaseBundle(releaseBundle *ReleaseBundleBody) error {
+func (cbs *CreateReleaseBundleService) execCreateReleaseBundle(releaseBundle *CreateReleaseBundleBody) error {
 	httpClientsDetails := cbs.DistDetails.CreateHttpClientDetails()
 	content, err := json.Marshal(releaseBundle)
 	if err != nil {
@@ -98,45 +62,8 @@ func (cbs *CreateReleaseBundleService) execCreateReleaseBundle(releaseBundle *Re
 	return errorutils.CheckError(err)
 }
 
-type ReleaseBundleBody struct {
-	Name              string        `json:"name"`
-	Version           string        `json:"version"`
-	DryRun            bool          `json:"dry_run"`
-	SignImmediately   bool          `json:"sign_immediately,omitempty"`
-	StoringRepository string        `json:"storing_repository,omitempty"`
-	Description       string        `json:"description,omitempty"`
-	ReleaseNotes      *ReleaseNotes `json:"release_notes,omitempty"`
-	BundleSpec        BundleSpec    `json:"spec"`
-}
-
-type ReleaseNotes struct {
-	Syntax  ReleaseNotesSyntax `json:"syntax,omitempty"`
-	Content string             `json:"content,omitempty"`
-}
-
-type BundleSpec struct {
-	Queries []BundleQuery `json:"queries"`
-}
-
-type BundleQuery struct {
-	QueryName string `json:"query_name,omitempty"`
-	Aql       string `json:"aql"`
-}
-
-type CreateReleaseBundleParams struct {
-	SpecFiles          []*artifactoryUtils.ArtifactoryCommonParams
-	Name               string
-	Version            string
-	SignImmediately    bool
-	StoringRepository  string
-	Description        string
-	ReleaseNotes       string
-	ReleaseNotesSyntax ReleaseNotesSyntax
-}
-
-func NewCreateBundleParams(name, version string) CreateReleaseBundleParams {
-	return CreateReleaseBundleParams{
-		Name:    name,
-		Version: version,
-	}
+type CreateReleaseBundleBody struct {
+	Name    string `json:"name"`
+	Version string `json:"version"`
+	ReleaseBundleBody
 }
